# Optimize `scanr` for improved performance