@usefulmove 2024-04-02
Yes, of course. I'll do that today.

- Duane -

On Tue, Apr 2, 2024 at 7:36 AM Conor Hoekstra ***@***.***>
wrote:

> @usefulmove <https://github.com/usefulmove> This looks great! Are you
> able to rebase on master / target it to master? It looks like you are
> targeting a "open-issues" branch from long time ago.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/codereport/racket-algorithms/pull/29#issuecomment-2032216950>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APO5WT2A6P2QO7FMBGBMRSLY3K67HAVCNFSM6AAAAABFSRSVNGVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDAMZSGIYTMOJVGA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>